Plane: move AP_Vehicle functions to dedicated file
