+++ ArduPlane/Plane.h
@@ -1115,6 +1115,8 @@ private:
 
 public:
     void failsafe_check(void);
+
+    // AP_Vehicle.h
     bool set_target_location(const Location& target_loc) override;
     bool get_target_location(Location& target_loc) override;
 };
